Normative: Clarify sign of modulo operation in GetISOPartsFromEpoch
Unlike the `%` operator, the result of the modulo operation in the spec
text has the sign of the right-hand operand, not the left-hand one. So
the current description of GetISOPartsFromEpoch was incorrect, and would
lead to incorrect results in some cases, and failed assertions in other
cases.

Add tests that will throw if this is implemented incorrectly.

(There are two other uses of modulo in this abstract operation, which I
believe are already correct.)